spam-message-detection/requirements.txt
alembic==1.7.5
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.4
certifi==2022.9.24
certipy==0.1.3
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
filelock==3.6.0
fonttools==4.28.2
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-telemetry==0.1.0
jupyterhub==1.5.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.7
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2022.3.28
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
pydub==0.25.1
Pygments==2.10.0
Pympler==0.9
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.10.31
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
selenium==4.0.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
SpeechRecognition==3.8.1
SQLAlchemy==1.4.27
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.14.0
virtualenv-clone==0.5.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wordcloud==1.8.2.2
wsproto==1.0.0